+++ src/main/authorization-response/Opts.ts
@@ -3,9 +3,9 @@ import { isExternalSignature, isExternalVerification, isInternalSignature, isInt
 import { AuthorizationResponseOpts, VerifyAuthorizationResponseOpts } from './types';
 
 export const assertValidResponseOpts = (opts: AuthorizationResponseOpts) => {
-  if (!opts /*|| !opts.redirectUri*/ || !opts.signatureType /*|| !opts.nonce*/ /* || !opts.did*/) {
+  if (!opts /*|| !opts.redirectUri*/ || !opts.signature /*|| !opts.nonce*/ /* || !opts.did*/) {
     throw new Error(SIOPErrors.BAD_PARAMS);
-  } else if (!(isInternalSignature(opts.signatureType) || isExternalSignature(opts.signatureType) || isSuppliedSignature(opts.signatureType))) {
+  } else if (!(isInternalSignature(opts.signature) || isExternalSignature(opts.signature) || isSuppliedSignature(opts.signature))) {
     throw new Error(SIOPErrors.SIGNATURE_OBJECT_TYPE_NOT_SET);
   }
 };