+++ src/main/request-object/types.ts
@@ -3,5 +3,5 @@ import { ExternalSignature, InternalSignature, NoSignature, ObjectBy, SuppliedSi
 
 export interface RequestObjectOpts<CT extends ClaimPayloadCommonOpts> extends ObjectBy {
   payload?: RequestObjectPayloadOpts<CT>; // for pass by value
-  signatureType: InternalSignature | ExternalSignature | SuppliedSignature | NoSignature; // Whether no signature is being used, internal (access to private key), or external (hosted using authentication), or supplied (callback supplied)
+  signature: InternalSignature | ExternalSignature | SuppliedSignature | NoSignature; // Whether no withSignature is being used, internal (access to private key), or external (hosted using authentication), or supplied (callback supplied)
 }